server/requirements.txt
aiofiles==24.1.0
alembic==1.15.2
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
autogen-agentchat==0.4.9.3
autogen-core==0.4.9.3
autogen-ext==0.4.9.3
autogenstudio==0.4.2.1
azure-ai-documentintelligence==1.0.2
azure-ai-inference==1.0.0b9
azure-core==1.34.0
azure-identity==1.23.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cobble==0.1.4
cryptography==44.0.3
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.2.2
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
jsonref==1.1.0
loguru==0.7.3
lxml==5.4.0
Mako==1.3.10
mammoth==1.9.0
markdown-it-py==3.0.0
markdownify==1.1.0
markitdown==0.0.2
MarkupSafe==3.0.2
mdurl==0.1.2
msal==1.32.3
msal-extensions==1.3.1
numpy==2.2.5
olefile==0.47
openai==1.78.1
openpyxl==3.1.5
opentelemetry-api==1.33.0
pandas==2.2.3
pathvalidate==3.2.3
pdfminer.six==20250506
pillow==11.2.1
playwright==1.52.0
protobuf==5.29.4
psycopg==3.2.9
puremagic==1.29
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.3
youtube-transcript-api==1.0.3
zipp==3.21.0
